+++ scene.py
@@ -1,1 +1,136 @@
-from set import Gc
+from Gctool import Gc
+indoor=Gc(
+'''
+Atlas 家庭內部....
+［序章：悲歌與艷陽交織的那場盛夏］
+
+那是一個炎熱得讓人頭暈目眩的午後。
+柏油路面被太陽烤得扭曲變形，空氣中瀰漫著一股乾燥的塵土味。
+我擦了擦額頭上的汗水，抬頭看向那棟熟悉的公寓。
+
+「這就是...前輩住的地方嗎？」
+懷著一種莫名忐忑的心情，我爬上了樓梯。
+明明只是來玩個電動、聊聊天，為什麼心跳會這麼快呢？
+也許是因為...今天的太陽實在太大了。
+扣扣扣扣.....
+提示：這是你唯一的反悔機會，請仔細思考決定
+
+r:進入
+l:光速離開
+'''
+)
+norun=Gc('''
+野獸先輩
+「來都來了...學弟啊，給個面子罷。」
+r:好！.
+l:好吧...
+''')
+house=Gc('''
+
+野獸先輩
+「這麼熱的天氣，真是辛苦你了啊。」
+「來，隨便坐吧，不用客氣。」
+{r}
+你
+「打...打擾了。」
+我拘謹地坐在那張看起來有點歷史的沙發上。
+房間裡雖然開著冷氣，但不知道為什麼，總覺得有一股燥熱感揮之不去。
+還有...空氣中似乎飄散著一股...讓人說不上來的味道？
+野獸先輩
+「外面很熱對吧？36度甚至38度都有可能呢。」
+你
+「是啊，真的快融化了...」
+野獸先輩
+「那你一定很渴了吧？」
+「稍微等一下喔...」
+先輩背對著我，肩膀微微聳動，似乎在調製著什麼。
+不知為何，看著那個背影，我突然有一種想逃跑的衝動。
+這是一種...野生動物的直覺嗎？
+野獸先輩
+「雖然只有紅茶...不過很好喝喔。」
+野獸先輩
+「來，請用。」
+「可是我... 特製 的喔。」
+r:大口喝下並稱讚
+l:小口啜飲並稱讚
+'''
+ )
+ch2=Gc('''
+[第二章：扁舟隨風起，赤子夜不歸］
+野獸先輩
+「{r} {l}」
+盛情難卻。
+我看著前輩那張充滿期待（？）的臉，實在不好意思拒絕。
+杯子裡的液體呈現出一種深沉的琥珀色，看起來只是一杯普通的紅茶。
+你
+「那我就不客氣了...」
+冰涼的液體滑過喉嚨，瞬間驅散了暑氣。
+味道...嗯？好像有點甜？又好像有點苦？
+說不上來，是一種從未喝過的味道。
+野獸先輩
+「怎麼樣？」
+你
+「呃...挺好喝的。」
+野獸先輩
+「是嗎？那就好...」
+「那就好啊...（小聲）」
+奇怪...
+怎麼突然覺得...眼皮好重...？
+明明剛喝了冰紅茶，身體卻變得像火燒一樣熱。
+從指尖開始，力氣彷彿被什麼東西抽走了一樣。
+「前...前輩...？」
+「我怎麼覺得...有點...」
+野獸先輩
+「怎麼了？臉色不太好喔？」
+前輩的身影在我眼中變成了兩個、三個...
+他站了起來。
+那個原本看起來不高的身型，此刻在我的視角裡，卻像是一座壓迫感十足的高山。
+野獸先輩
+「沒事吧？是不是中暑了？」
+「來，去那邊躺一下吧。」
+「不要客氣... 真的不要客氣喔。」
+野獸先輩
+「哼哼...」
+
+
+
+你dady sksbsjajsbjsanababhaajab我曾說sksndjsjsjjasnsajananajnaanamamamam你snnssjdjsjsnjaaksjsnjwj蘿莉
+l:堅守底線 奮力掙扎
+r:放棄希望
+'''
+)
+end=Gc('''
+
+[終章：終末的囈語］
+
+四十度的熱氣，也無法勸退野獸先輩激烈的活動
+你渾身刺痛的躺在床上，眼淚一滴一滴的掉了下來...
+真可惜我是睡著的啊...
+（the end)
+
+
+'''
+)
+good_end=Gc('''
+［終章：彼方之矢］
+你帶著享受的心情離開了先輩的家
+只有你知道，一切都是你計劃好的
+幾個月前，你就一點一點的下藥
+一步一步的接近先輩
+月色緩緩流過你輕盈的身影
+最終，你的腳步發散在蟬鳴的間隙
+只留下一句囈語
+「下一個...誰呢？」
+（the end)
+
+''' )
+
+ch2.spec(l="你喝東西的樣子真是雷普啊（震聲）",r="非常好")
+house.spec(r="不過我看你挺精神啊 哈哈")
+
+ch1=Gc.Layer(
+        indoor,
+        (house, norun),
+        ch2,
+        (end, good_end),
+    )
